// packages/fasset-indexer-core/chain/typechain/AssetManager.ts
export declare namespace AgentInfo {
  export type InfoStruct = {
    status: BigNumberish;
    ownerManagementAddress: AddressLike;
    ownerWorkAddress: AddressLike;
    collateralPool: AddressLike;
    underlyingAddressString: string;
    publiclyAvailable: boolean;
    feeBIPS: BigNumberish;
    poolFeeShareBIPS: BigNumberish;
    vaultCollateralToken: AddressLike;
    mintingVaultCollateralRatioBIPS: BigNumberish;
    mintingPoolCollateralRatioBIPS: BigNumberish;
    freeCollateralLots: BigNumberish;
    totalVaultCollateralWei: BigNumberish;
    freeVaultCollateralWei: BigNumberish;
    vaultCollateralRatioBIPS: BigNumberish;
    totalPoolCollateralNATWei: BigNumberish;
    freePoolCollateralNATWei: BigNumberish;
    poolCollateralRatioBIPS: BigNumberish;
    totalAgentPoolTokensWei: BigNumberish;
    announcedVaultCollateralWithdrawalWei: BigNumberish;
    announcedPoolTokensWithdrawalWei: BigNumberish;
    freeAgentPoolTokensWei: BigNumberish;
    mintedUBA: BigNumberish;
    reservedUBA: BigNumberish;
    redeemingUBA: BigNumberish;
    poolRedeemingUBA: BigNumberish;
    dustUBA: BigNumberish;
    ccbStartTimestamp: BigNumberish;
    liquidationStartTimestamp: BigNumberish;
    maxLiquidationAmountUBA: BigNumberish;
    liquidationPaymentFactorVaultBIPS: BigNumberish;
    liquidationPaymentFactorPoolBIPS: BigNumberish;
    underlyingBalanceUBA: BigNumberish;
    requiredUnderlyingBalanceUBA: BigNumberish;
    freeUnderlyingBalanceUBA: BigNumberish;
    announcedUnderlyingWithdrawalId: BigNumberish;
    buyFAssetByAgentFactorBIPS: BigNumberish;
    poolExitCollateralRatioBIPS: BigNumberish;
    poolTopupCollateralRatioBIPS: BigNumberish;
    poolTopupTokenPriceFactorBIPS: BigNumberish;
  };

  export type InfoStructOutput = [
    status: bigint,
    ownerManagementAddress: string,
    ownerWorkAddress: string,
    collateralPool: string,
    underlyingAddressString: string,
    publiclyAvailable: boolean,
    feeBIPS: bigint,
    poolFeeShareBIPS: bigint,
    vaultCollateralToken: string,
    mintingVaultCollateralRatioBIPS: bigint,
    mintingPoolCollateralRatioBIPS: bigint,
    freeCollateralLots: bigint,
    totalVaultCollateralWei: bigint,
    freeVaultCollateralWei: bigint,
    vaultCollateralRatioBIPS: bigint,
    totalPoolCollateralNATWei: bigint,
    freePoolCollateralNATWei: bigint,
    poolCollateralRatioBIPS: bigint,
    totalAgentPoolTokensWei: bigint,
    announcedVaultCollateralWithdrawalWei: bigint,
    announcedPoolTokensWithdrawalWei: bigint,
    freeAgentPoolTokensWei: bigint,
    mintedUBA: bigint,
    reservedUBA: bigint,
    redeemingUBA: bigint,
    poolRedeemingUBA: bigint,
    dustUBA: bigint,
    ccbStartTimestamp: bigint,
    liquidationStartTimestamp: bigint,
    maxLiquidationAmountUBA: bigint,
    liquidationPaymentFactorVaultBIPS: bigint,
    liquidationPaymentFactorPoolBIPS: bigint,
    underlyingBalanceUBA: bigint,
    requiredUnderlyingBalanceUBA: bigint,
    freeUnderlyingBalanceUBA: bigint,
    announcedUnderlyingWithdrawalId: bigint,
    buyFAssetByAgentFactorBIPS: bigint,
    poolExitCollateralRatioBIPS: bigint,
    poolTopupCollateralRatioBIPS: bigint,
    poolTopupTokenPriceFactorBIPS: bigint
  ] & {
    status: bigint;
    ownerManagementAddress: string;
    ownerWorkAddress: string;
    collateralPool: string;
    underlyingAddressString: string;
    publiclyAvailable: boolean;
    feeBIPS: bigint;
    poolFeeShareBIPS: bigint;
    vaultCollateralToken: string;
    mintingVaultCollateralRatioBIPS: bigint;
    mintingPoolCollateralRatioBIPS: bigint;
    freeCollateralLots: bigint;
    totalVaultCollateralWei: bigint;
    freeVaultCollateralWei: bigint;
    vaultCollateralRatioBIPS: bigint;
    totalPoolCollateralNATWei: bigint;
    freePoolCollateralNATWei: bigint;
    poolCollateralRatioBIPS: bigint;
    totalAgentPoolTokensWei: bigint;
    announcedVaultCollateralWithdrawalWei: bigint;
    announcedPoolTokensWithdrawalWei: bigint;
    freeAgentPoolTokensWei: bigint;
    mintedUBA: bigint;
    reservedUBA: bigint;
    redeemingUBA: bigint;
    poolRedeemingUBA: bigint;
    dustUBA: bigint;
    ccbStartTimestamp: bigint;
    liquidationStartTimestamp: bigint;
    maxLiquidationAmountUBA: bigint;
    liquidationPaymentFactorVaultBIPS: bigint;
    liquidationPaymentFactorPoolBIPS: bigint;
    underlyingBalanceUBA: bigint;
    requiredUnderlyingBalanceUBA: bigint;
    freeUnderlyingBalanceUBA: bigint;
    announcedUnderlyingWithdrawalId: bigint;
    buyFAssetByAgentFactorBIPS: bigint;
    poolExitCollateralRatioBIPS: bigint;
    poolTopupCollateralRatioBIPS: bigint;
    poolTopupTokenPriceFactorBIPS: bigint;
  };
}